biaoxingcaiji/DjangoTest/pyproject.toml
[project]
name = "phenotype-system"
version = "0.1.0"
description = "表型采集系统 - Django REST Framework 项目"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "django>=4.2,<5.0",
    "pillow>=10.0.0",
    "mysqlclient>=2.2.0",
    "djangorestframework>=3.14.0",
    "django-filter>=23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"